% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task_functions.R
\name{annotate_cell_types}
\alias{annotate_cell_types}
\title{Annotate cell types using SingleR}
\usage{
annotate_cell_types(seurat_obj, ref, labels)
}
\arguments{
\item{seurat_obj}{A Seurat object.}

\item{ref}{Reference matrix (genes × cell types).}

\item{labels}{Cell type labels corresponding to columns in "ref".}
}
\value{
A list with Seurat object with cell type metadata and cell type UMAP plot labeled by cell type.
}
\description{
Uses SingleR to assign cell type labels to each cell based on a reference.
}
\examples{
result <- annotate_cell_types(seurat_obj, ref_matrix, ref_labels)
}
