% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task_functions.R
\name{extract_gene_annotations_from_gtf}
\alias{extract_gene_annotations_from_gtf}
\title{Extract gene annotations from the GTF file you want to analyze}
\usage{
extract_gene_annotations_from_gtf(gtf_path)
}
\arguments{
\item{gtf_path}{Path to the GTF annotation file.}
}
\value{
A data.frame with gene_id and gene_type of filtered elements annotated as genes
}
\description{
Extracts gene_id and gene_type from a GTF file, keeping only entries where the feature type is "gene".
}
\details{
"dplyr" package required
}
\examples{
gene_info <- extract_gene_annotations_from_gtf("Homo_sapiens.GRCh38.111.gtf")
}
