% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task_functions.R
\name{filter_protein_coding_genes}
\alias{filter_protein_coding_genes}
\title{Filter Seurat object to keep only protein-coding genes}
\usage{
filter_protein_coding_genes(seurat_obj, gtf_path)
}
\arguments{
\item{seurat_obj}{A Seurat object.}

\item{gtf_path}{Path to the GTF annotation file.}
}
\value{
A filtered Seurat object containing only protein-coding genes.
}
\description{
This function reads a GTF file, extracts gene_id and gene_type using extract_attribute,
filters for protein-coding genes, and subsets the Seurat object accordingly.
}
\details{
"dplyr" package required
}
\examples{
seurat_pc <- filter_protein_coding_genes(seurat_obj, "Homo_sapiens.GRCh38.111.gtf")
}
