% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task_functions.R
\name{plot_gene_expression_summary}
\alias{plot_gene_expression_summary}
\title{Plot number of genes with >=3 UMIs per cell}
\usage{
plot_gene_expression_summary(seurat_obj)
}
\arguments{
\item{seurat_obj}{A Seurat object.}
}
\value{
A list with Seurat object with metadata column `genes_expressed_3UMI, and a violin plot.
}
\description{
Plot number of genes with >=3 UMIs per cell
}
\examples{
result <- plot_gene_expression_summary(seurat_obj)
result$plot
}
