% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task_functions.R
\name{run_umap_custom}
\alias{run_umap_custom}
\title{Run UMAP using selected principal componenet(PCs)}
\usage{
run_umap_custom(seurat_obj, dims)
}
\arguments{
\item{seurat_obj}{A Seurat object.}

\item{dims}{PCs to use, e.g. 1:10.}
}
\value{
A list with Seurat object and UMAP plot object with cluster labels.
}
\description{
Executes UMAP dimensionality reduction using user-defined principal components.
}
\examples{
result <- run_umap_custom(seurat_obj, dims = 1:10)
}
